node-main/deps/v8/test/webkit/fast/js/Promise-then-expected.txt

Test Promise.prototype.then

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInInit is undefined
PASS firstPromise instanceof Promise is true
PASS secondPromise instanceof Promise is true
PASS thisInOnFulfilled is undefined
PASS result is "hello"
PASS result is "world"
PASS rejected
PASS result is "exception"
PASS resolved
PASS successfullyParsed is true

TEST COMPLETE
